Testing new HillCoordinate syntax
Fixed the call to HillCoordinate.dx to allow callable parameters. Added exception to HillCoordinate.eq_interval to handle the case where ell or delta is a variable parameter for a HillComponent. In this case the variable parameter will not occur as an attribute for the HillComponent and should throw and exception with appropriate message.