Added overloaded derivatives to ToggleSwitch
Overloaded the .dx2diff, .diff2, and .dxdiff2 methods for the
ToggleSwitch. This requires identifying all Hill coefficients as a
single variable parameter. The general formula is universal so these
derivatives are the prototypical example for how to identify Hill
coefficients in general. In the future it may make sense to define
a HillModel subclass which has this identification.

Note: The universal Hill coefficient is always a variable parameter
in this class and it is always passed as the first parameter.