Added example of hill coefficient minimization and DSGRN fiber sampling
hill_minima_toggleswitch includes an example of optimization along a saddle-node
bifurcation surface for the ToggleSwitch model. In this example the Loss function is
either the hill coefficient, or the hill coefficient divided by the parameter distance to
a linear subspace of degenerate parameters for which the hill coefficient is equal to 1.

toggle_switch_fiber_sample is a script for sampling and analyzing the distribution
of saddle node points in parameter space. This example looks at the fiber of the
upper right corner of the central DSGRN parameter region projected into 3 dimensional
space. This fiber is characterized by the equations:

gamma_1 = theta_1 = theta_2 = 1  (via non-dimensionalization)
ell_1/gamma_1 = theta_1 = 1 (horizontal boundary condition)
ell_2/gamma_2 = theta_2 = 1 (vertical boundary condition)

The parameters of interest for each point in this fiber are the triples:
(delta_1, gamma_2, delta_2)