Debugging after refactorization and identifying Hill coefficients in EMT model
Added identification of Hill coefficients for the EMT model. This reduces the number of parameters from 54 to 43. This has also revealed some bugs in the HillModel class which are being resolved. In addition, I added a decorator function "verify_call" to the HillCont library. It can decorate any evaluation method for HillCoordinates or HillModels which have inputs of the form (self, x, *parameter, **kwargs). The decorator checks that the size of x and all passed parameters agree with the expected state and parameter dimensions of the HillCoordinate or HillModel.

The decorator has been added to all evaluation methods (function calls and derivatives) and can be removed once the refactored code has passed thorough testing.