Added selective derivative indexing to the EMT derivative methods.
This is required in order for the SaddleNode class methods to work properly. This implementation is for identified Hill coefficients and the code is similar to the implementation for the Toggle Switch (with identified Hill coefficients). Selective indexing (and therefore the SaddleNode class) is still not functional in the general HillModel class.